# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "v1/webhooks": {
            "post": {
                "tags": [
                    "Company"
                ],
                "description": "Add webhook to comapny",
                "operationId": "addWebhook",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "url",
                        "in": "formData",
                        "description": "Webhook url",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "code",
                        "in": "formData",
                        "description": "Webhook code",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "company_id",
                        "in": "formData",
                        "description": "Company ID. Later we may will not need this field, it may replaced by company authentication key",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation returns success message",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Errors",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Fields are invalid"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Server Error"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Driver Profile",
                "description": "Get driver profile information",
                "operationId": "edit",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "driver_id",
                        "in": "query",
                        "description": "Driver ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation returns success message",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Errors",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Fields are invalid"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Driver not found",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Server Error"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/v1/profile/edit": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update Driver Profile",
                "description": "Update Driver Profile Info",
                "operationId": "edit",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "driver_id",
                        "in": "formData",
                        "description": "Driver ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Driver Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "Driver Username",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "description": "Driver Image",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "mobile",
                        "in": "formData",
                        "description": "Driver Mobile",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation returns success message",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Errors",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Fields are invalid"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Driver not found",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Server Error"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}